refactor(writers): share the line-fill engine across writers
Lift the greedy line-filling engine — inline pieces, fill/fill_offset, column-width measurement, list-tightness, and ordered-list delimiter wrapping — out of the plain writer into the shared writers module, so the plain, HTML, and LaTeX writers all build on one implementation instead of carrying their own copies.
